+++ LayerKit.embeddedframework/LayerKit.framework/Versions/A/Resources/layer-client-messaging-schema.bundle/layer-client-messaging-schema.sql
@@ -25,7 +25,7 @@ CREATE TABLE "conversations" (
   deleted_at DATETIME,
   object_identifier TEXT UNIQUE NOT NULL,
   version INT NOT NULL
-, has_unread_messages INTEGER NOT NULL DEFAULT 0);
+, has_unread_messages INTEGER NOT NULL DEFAULT 0, is_distinct INTEGER NOT NULL DEFAULT 0);
 
 CREATE TABLE "deleted_message_parts" (
   database_identifier INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT,
@@ -54,8 +54,7 @@ CREATE TABLE "events" (
   member_id TEXT,
   target_seq INTEGER,
   stream_database_identifier INTEGER NOT NULL,
-  version INTEGER, 
-  client_id TEXT,
+  client_id BLOB, creator_name TEXT,
   UNIQUE(stream_database_identifier, seq),
   FOREIGN KEY(stream_database_identifier) REFERENCES streams(database_identifier) ON DELETE CASCADE
 );
@@ -104,13 +103,13 @@ CREATE TABLE "messages" (
   sent_at DATETIME,
   received_at DATETIME,
   deleted_at DATETIME,
-  user_id TEXT NOT NULL,
+  user_id TEXT,
   seq INTEGER,
   conversation_database_identifier INTEGER NOT NULL,
   event_database_identifier INTEGER UNIQUE,
   version INTEGER NOT NULL,
   object_identifier TEXT UNIQUE NOT NULL,
-  message_index INTEGER, is_unread INTEGER NOT NULL DEFAULT 0,
+  message_index INTEGER, is_unread INTEGER NOT NULL DEFAULT 0, user_name TEXT,
   UNIQUE(conversation_database_identifier, seq),
   FOREIGN KEY(conversation_database_identifier) REFERENCES conversations(database_identifier) ON DELETE CASCADE,
   FOREIGN KEY(event_database_identifier) REFERENCES events(database_identifier) ON DELETE CASCADE
@@ -147,11 +146,11 @@ CREATE TABLE "streams" (
   stream_id BLOB UNIQUE,
   seq INTEGER NOT NULL DEFAULT 0,
   client_seq INTEGER NOT NULL DEFAULT 0,
-  version INTEGER, 
-  client_id TEXT, 
+  client_id BLOB, 
   deleted_at DATETIME, 
   min_synced_seq INTEGER, 
-  max_synced_seq INTEGER, metadata_timestamp INTEGER);
+  max_synced_seq INTEGER, metadata_timestamp INTEGER, is_distinct INTEGER NOT NULL DEFAULT 0
+);
 
 CREATE TABLE syncable_changes (
   change_identifier INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT,
@@ -213,13 +212,13 @@ CREATE INDEX streams_deleted_at_idx ON streams(deleted_at);
 CREATE INDEX synced_changes_table_name_idx ON synced_changes(table_name);
 
 
-CREATE TRIGGER tombstone_duplicate_events_by_client_id 
+CREATE TRIGGER tombstone_duplicate_events_by_client_id
 AFTER INSERT ON events
 FOR EACH ROW WHEN NEW.client_id IS NOT NULL
 BEGIN
-	UPDATE events SET type = 10 
-	WHERE database_identifier = NEW.database_identifier
-	AND (SELECT count(*) FROM events WHERE client_id = NEW.client_id) > 1;
+       UPDATE events SET type = 10
+       WHERE database_identifier = NEW.database_identifier
+       AND (SELECT count(*) FROM events WHERE client_id = NEW.client_id) > 1;
 END;
 
 CREATE TRIGGER track_deletes_of_conversation_participants AFTER UPDATE OF deleted_at ON conversation_participants
@@ -277,15 +276,15 @@ BEGIN
 END;
 
 CREATE TRIGGER track_inserts_of_events_delete
-AFTER INSERT ON events FOR EACH ROW 
+AFTER INSERT ON events FOR EACH ROW
 WHEN NEW.seq IS NOT NULL AND NEW.type = 9 AND NOT EXISTS (SELECT 1 FROM events WHERE client_id = NEW.client_id AND database_identifier != NEW.database_identifier)
 BEGIN
   INSERT OR IGNORE INTO synced_changes(table_name, row_identifier, change_type) VALUES ('events', (SELECT database_identifier FROM events WHERE stream_database_identifier = NEW.stream_database_identifier AND seq = NEW.target_seq), 2);
-  UPDATE events SET type = 10, subtype = NULL, external_content_id = NULL, member_id = NULL, target_seq = NULL, version = NULL WHERE database_identifier = (SELECT database_identifier FROM events WHERE stream_database_identifier = NEW.stream_database_identifier AND seq = NEW.target_seq);
+  UPDATE events SET type = 10, subtype = NULL, external_content_id = NULL, member_id = NULL, target_seq = NULL WHERE database_identifier = (SELECT database_identifier FROM events WHERE stream_database_identifier = NEW.stream_database_identifier AND seq = NEW.target_seq);
 END;
 
 CREATE TRIGGER track_inserts_of_events_non_delete
-AFTER INSERT ON events FOR EACH ROW 
+AFTER INSERT ON events FOR EACH ROW
 WHEN NEW.seq IS NOT NULL AND NEW.type != 9 AND NOT EXISTS (SELECT 1 FROM events WHERE client_id = NEW.client_id AND database_identifier != NEW.database_identifier)
 BEGIN
   INSERT INTO synced_changes(table_name, row_identifier, change_type) VALUES ('events', NEW.database_identifier, 0);
@@ -370,6 +369,12 @@ BEGIN
   INSERT INTO synced_changes(table_name, row_identifier, change_type) VALUES ('remote_keyed_values', NEW.database_identifier, 1);
 END;
 
+CREATE TRIGGER track_updates_of_stream_id_for_events AFTER UPDATE OF stream_database_identifier ON events
+WHEN NEW.stream_database_identifier IS NOT NULL AND OLD.stream_database_identifier IS NOT NULL AND NEW.stream_database_identifier != OLD.stream_database_identifier
+BEGIN
+  INSERT INTO synced_changes(table_name, row_identifier, change_type) VALUES ('events', NEW.database_identifier, 1);
+END;
+
 CREATE TRIGGER track_updates_of_streams AFTER UPDATE OF stream_id ON streams
 WHEN NEW.stream_id IS NOT NULL AND OLD.stream_id IS NULL
 BEGIN
@@ -443,3 +448,9 @@ INSERT INTO schema_migrations (version) VALUES (20150202170209118);
 INSERT INTO schema_migrations (version) VALUES (20150207191203003);
 
 INSERT INTO schema_migrations (version) VALUES (20150210133608257);
+
+INSERT INTO schema_migrations (version) VALUES (20150316180034638);
+
+INSERT INTO schema_migrations (version) VALUES (20150319131356212);
+
+INSERT INTO schema_migrations (version) VALUES (20150330135300206);